finishing to set up Globe on bentogrid
